# Fix type instability when passing type to a function, concretize default types