--- a/examples/constituents/jetreco-constituents-nb.jl
+++ b/examples/constituents/jetreco-constituents-nb.jl
@@ -32,7 +32,7 @@ cluster_seq = jet_reconstruct(events[event_no]; algorithm = JetAlgorithm.Kt, R =
 md"Retrieve the exclusive pj_jets, but as `PseudoJet` types"
 
 # ╔═╡ 0d8d4664-915f-4f28-9d5a-6e03cb8d7d8b
-pj_jets = inclusive_jets(cluster_seq; ptmin = 5.0, T = PseudoJet)
+pj_jets = inclusive_jets(cluster_seq, PseudoJet; ptmin = 5.0)
 
 # ╔═╡ 0bd764f9-d427-43fc-8342-603b6759ec8f
 md"Get the constituents of the first jet"