--- a/examples/substructure/jet-grooming.jl
+++ b/examples/substructure/jet-grooming.jl
@@ -9,7 +9,7 @@ events = read_final_state_particles(input_file)
 event_no = 1
 
 cluster_seq = jet_reconstruct(events[event_no]; algorithm = JetAlgorithm.CA, R = 1.0)
-jets = inclusive_jets(cluster_seq; ptmin = 5.0, T = PseudoJet)
+jets = inclusive_jets(cluster_seq, PseudoJet; ptmin = 5.0)
 
 filter = (radius = 0.3, hardest_jets = 3)