--- a/src/ClusterSequence.jl
+++ b/src/ClusterSequence.jl
@@ -215,16 +215,16 @@ function add_step_to_history!(clusterseq::ClusterSequence, parent1, parent2, jet
 end
 
 """
-    inclusive_jets(clusterseq::ClusterSequence{U}; ptmin = 0.0, T = LorentzVectorCyl) where {U}
+    inclusive_jets(clusterseq::ClusterSequence{U}, ::Type{T} = LorentzVectorCyl; ptmin = 0.0) where {T, U}
 
 Return all inclusive jets of a ClusterSequence with pt > ptmin.
 
 # Arguments
 - `clusterseq::ClusterSequence`: The `ClusterSequence` object containing the
   clustering history and jets.
+- `::Type{T} = LorentzVectorCyl`: The return type used for the selected jets.
 - `ptmin::Float64 = 0.0`: The minimum transverse momentum (pt) threshold for the
   inclusive jets.
-- `T = LorentzVectorCyl`: The return type used for the selected jets.
 
 # Returns
 An array of `T` objects representing the inclusive jets.
@@ -245,8 +245,8 @@ currently unrecognised types will return `LorentzVectorCyl`).
 inclusive_jets(clusterseq; ptmin = 10.0)
 ```
 """
-function inclusive_jets(clusterseq::ClusterSequence{U}; ptmin = 0.0,
-                        T = LorentzVectorCyl) where {U}
+function inclusive_jets(clusterseq::ClusterSequence{U}, ::Type{T} = LorentzVectorCyl;
+                        ptmin = 0.0) where {T, U}
     pt2min = ptmin * ptmin
     jets_local = T[]
     # sizehint!(jets_local, length(clusterseq.jets))
@@ -272,7 +272,7 @@ function inclusive_jets(clusterseq::ClusterSequence{U}; ptmin = 0.0,
 end
 
 """
-    exclusive_jets(clusterseq::ClusterSequence{U}; dcut = nothing, njets = nothing, T = LorentzVectorCyl) where {U}
+    exclusive_jets(clusterseq::ClusterSequence{U}, ::Type{T} = LorentzVectorCyl; dcut = nothing, njets = nothing) where {T, U}
 
 Return all exclusive jets of a ClusterSequence, with either a specific number of
 jets or a cut on the maximum distance parameter.
@@ -280,13 +280,13 @@ jets or a cut on the maximum distance parameter.
 # Arguments
 - `clusterseq::ClusterSequence`: The `ClusterSequence` object containing the
   clustering history and jets.
+- `::Type{T} = LorentzVectorCyl`: The return type used for the selected jets.
 - `dcut::Union{Nothing, Real}`: The distance parameter used to define the
   exclusive jets. If `dcut` is provided, the number of exclusive jets will be
   calculated based on this parameter.
 - `njets::Union{Nothing, Integer}`: The number of exclusive jets to be
   calculated. If `njets` is provided, the distance parameter `dcut` will be
   calculated based on this number.
-- `T = LorentzVectorCyl`: The return type used for the selected jets.
 
 **Note**: Either `dcut` or `njets` must be provided (but not both).
 
@@ -308,11 +308,11 @@ currently unrecognised types will return `LorentzVectorCyl`).
 # Examples
 ```julia
 exclusive_jets(clusterseq, dcut = 20.0)
-exclusive_jets(clusterseq, njets = 3, T = PseudoJet)
+exclusive_jets(clusterseq, PseudoJet, njets = 3)
 ```
 """
-function exclusive_jets(clusterseq::ClusterSequence{U}; dcut = nothing, njets = nothing,
-                        T = LorentzVectorCyl) where {U}
+function exclusive_jets(clusterseq::ClusterSequence{U}, ::Type{T} = LorentzVectorCyl; dcut = nothing, njets = nothing,
+                        ) where {T, U}
     if isnothing(dcut) && isnothing(njets)
         throw(ArgumentError("Must pass either a dcut or an njets value"))
     end