--- a/test/data/hepmc32summary.jl
+++ b/test/data/hepmc32summary.jl
@@ -35,8 +35,9 @@ function main()
     args = parse_command_line(ARGS)
 
     for file in args[:files]
-        events = read_final_state_particles(file; maxevents = args[:maxevents],
-                                            skipevents = args[:skip], T = LorentzVector)
+        events = read_final_state_particles(file, LorentzVector;
+                                            maxevents = args[:maxevents],
+                                            skipevents = args[:skip])
         n_events = length(events)
         n_particles = Int[]
         for e in events