--- a/test/test-compare-types.jl
+++ b/test/test-compare-types.jl
@@ -32,8 +32,8 @@ function do_test_compare_types(strategy::RecoStrategy.Strategy;
     end
 
     # From LorentzVector
-    events_lv::Vector{Vector{LorentzVector}} = read_final_state_particles(events_file_pp;
-                                                                          T = LorentzVector)
+    events_lv::Vector{Vector{LorentzVector}} = read_final_state_particles(events_file_pp,
+                                                                          LorentzVector)
     jet_collection_lv = FinalJets[]
     for (ievt, event) in enumerate(events_lv)
         finaljets = final_jets(inclusive_jets(jet_reconstruction(event, R = distance,