As described in #185, passing a type as a keyword argument to a function causes a type instability. Passing type as a positional argument is fine. [reference](https://docs.julialang.org/en/v1/manual/methods/#Note-on-Optional-and-keyword-Arguments)

I propose here to change keyword argument `T=DefaultT` in `inclusive_jets`, `exclusive_jets`, etc. to be the last positional argument (and also constrain it to be `Type`), This eliminates the type instability but changes the interface. The last positional argument is a different convention that `Base.convert` and other standard functions but they don't allow the default value which we want to have

Another thing as pointed by @Moelf  is that parametric types used by in the default values (or used in examples) could be concretized.

Performance wise I don't see any noticeable difference with `julia --project instrumented-jetreco.jl --algorithm=AntiKt -R 0.4 ../test/data/events.pp13TeV.hepmc3.zst -m 32` due to the type instability

- main https://github.com/JuliaHEP/JetReconstruction.jl/blob/24631147985e1eb11a454f07a86ada653f54b50d/examples/instrumented-jetreco.jl#L166
  ```
  Processed 100 events 32 times
  Average time per event 128.2739265625 ± 3.2664981607124073 μs
  Lowest time per event 122.77002 μs
  ```
- main instrumented-jetreco.jl#L166 changed to `inclusive_jets(cs; ptmin = ptmin, T=LorentzVectorCyl)` (should be the worst)
  ```
  Processed 100 events 32 times
  Average time per event 128.9642315625 ± 3.379545577028563 μs
  Lowest time per event 122.93741 μs
  ```
- this PR with instrumented-jetreco.jl#L166 changed to `inclusive_jets(cs, LorentzVectoCyl; ptmin = ptmin)`
  ```
  Processed 100 events 32 times
  Average time per event 127.5241915625 ± 2.8969402308373184 μs
  Lowest time per event 122.15337 μs
  ```

- this PR with instrumented-jetreco.jl#L166 changed to `inclusive_jets(cs, LorentzVectorCyl{Float64}; ptmin = ptmin)`
  ```
  Processed 100 events 32 times
  Average time per event 126.56505062500004 ± 3.578129031677597 μs
  Lowest time per event 120.3919 μs
  ```

Also `preprocess_ptscheme` had wrong name in docstring and wasn't using `jet_type` argument

Fixes #185
Closes #186