@graeme-a-stewart 2025-07-20
This is great! Thanks @m-fila. I had missed the issue with _named_ arguments back when I added the type option to the jet selection functions. It's very useful that you picked this up with the pre-compilation work on the package.